@@ -1,34 +1,34 @@
-part of 'job_card_bloc.dart';
+// part of 'job_card_bloc.dart';
 
-@immutable
-abstract class JobCardHeaderEvent {}
+// @immutable
+// abstract class JobCardHeaderEvent {}
 
-class GetInProgressByDateHeader extends JobCardHeaderEvent {
-  GetInProgressByDateHeader({@required this.date});
-  final int date;
-}
+// class GetInProgressByDateHeader extends JobCardHeaderEvent {
+//   GetInProgressByDateHeader({@required this.date});
+//   final int date;
+// }
 
-class GetInProgressByAllHeader extends JobCardHeaderEvent {}
+// class GetInProgressByAllHeader extends JobCardHeaderEvent {}
 
-class GetInProgressByMonthHeader extends JobCardHeaderEvent {
-  GetInProgressByMonthHeader({@required this.month});
-  final int month;
-}
+// class GetInProgressByMonthHeader extends JobCardHeaderEvent {
+//   GetInProgressByMonthHeader({@required this.month});
+//   final int month;
+// }
 
-class GetInProgressByWeekHeader extends JobCardHeaderEvent {
-  GetInProgressByWeekHeader({@required this.week});
-  final int week;
-}
+// class GetInProgressByWeekHeader extends JobCardHeaderEvent {
+//   GetInProgressByWeekHeader({@required this.week});
+//   final int week;
+// }
 
-class GetInProgressByYearHeader extends JobCardHeaderEvent {
-  GetInProgressByYearHeader({@required this.year});
-  final int year;
-}
+// class GetInProgressByYearHeader extends JobCardHeaderEvent {
+//   GetInProgressByYearHeader({@required this.year});
+//   final int year;
+// }
 
-abstract class JobCardBodyEvent {}
+// abstract class JobCardBodyEvent {}
 
-class GetInProgressBody extends JobCardBodyEvent {
-  GetInProgressBody({@required this.type, @required this.filterNumber});
-  final String type;
-  final int filterNumber;
-}
+// class GetInProgressBody extends JobCardBodyEvent {
+//   GetInProgressBody({@required this.type, @required this.filterNumber});
+//   final String type;
+//   final int filterNumber;
+// }
