@@ -1,45 +1,45 @@
-part of 'job_card_bloc.dart';
+// part of 'job_card_bloc.dart';
 
-@immutable
-abstract class JobCardHeaderState {}
+// @immutable
+// abstract class JobCardHeaderState {}
 
-class BuildHeaderByDate extends JobCardHeaderState {
-  BuildHeaderByDate({@required this.day});
-  final int day;
-}
+// class BuildHeaderByDate extends JobCardHeaderState {
+//   BuildHeaderByDate({@required this.day});
+//   final int day;
+// }
 
-class BuildHeaderByAll extends JobCardHeaderState {}
+// class BuildHeaderByAll extends JobCardHeaderState {}
 
-class BuildHeaderByWeek extends JobCardHeaderState {
-  BuildHeaderByWeek({@required this.week});
-  final int week;
-}
+// class BuildHeaderByWeek extends JobCardHeaderState {
+//   BuildHeaderByWeek({@required this.week});
+//   final int week;
+// }
 
-class BuildHeaderByMonth extends JobCardHeaderState {
-  BuildHeaderByMonth({@required this.month});
-  final int month;
-}
+// class BuildHeaderByMonth extends JobCardHeaderState {
+//   BuildHeaderByMonth({@required this.month});
+//   final int month;
+// }
 
-class BuildHeaderByYear extends JobCardHeaderState {
-  BuildHeaderByYear({@required this.year});
-  final int year;
-}
+// class BuildHeaderByYear extends JobCardHeaderState {
+//   BuildHeaderByYear({@required this.year});
+//   final int year;
+// }
 
-//................................................
+// //................................................
 
-abstract class JobCardBodyState {}
+// abstract class JobCardBodyState {}
 
-class BuildBodyLoading extends JobCardBodyState {}
+// class BuildBodyLoading extends JobCardBodyState {}
 
-class BuildBodyLoaded extends JobCardBodyState {
-  BuildBodyLoaded({@required this.jobCards});
-  final List<MyJobCard> jobCards;
-}
+// class BuildBodyLoaded extends JobCardBodyState {
+//   BuildBodyLoaded({@required this.jobCards});
+//   final List<MyJobCard> jobCards;
+// }
 
-// class BuildBodyByDate extends JobCardBodyState {}
+// // class BuildBodyByDate extends JobCardBodyState {}
 
-// class BuildBodyByWeek extends JobCardBodyState {}
+// // class BuildBodyByWeek extends JobCardBodyState {}
 
-// class BuildBodyByMonth extends JobCardBodyState {}
+// // class BuildBodyByMonth extends JobCardBodyState {}
 
-// class BuildBodyByYear extends JobCardBodyState {}
+// // class BuildBodyByYear extends JobCardBodyState {}
