@@ -1,11 +1,11 @@
-part of 'login_bloc.dart';
+// part of 'login_bloc.dart';
 
-@immutable
-abstract class LoginEvent {}
+// @immutable
+// abstract class LoginEvent {}
 
-class Login extends LoginEvent {
-  final String userName;
-  final String password;
+// class Login extends LoginEvent {
+//   final String userName;
+//   final String password;
 
-  Login({@required this.userName, @required this.password});
-}
+//   Login({@required this.userName, @required this.password});
+// }
