@@ -1,10 +1,10 @@
-part of 'landingpage_bloc.dart';
+// part of 'landingpage_bloc.dart';
 
-@immutable
-abstract class LandingPageState {}
+// @immutable
+// abstract class LandingPageState {}
 
-class LandingPageInitial extends LandingPageState {}
+// class LandingPageInitial extends LandingPageState {}
 
-class StartLoginPage extends LandingPageState {}
+// class StartLoginPage extends LandingPageState {}
 
-class StartHomePage extends LandingPageState {}
+// class StartHomePage extends LandingPageState {}
