@@ -1,5 +1,9 @@
+import 'package:airpmp_mobility/constants/enums.dart';
+import 'package:airpmp_mobility/screens_new/job_screen.dart';
 import 'package:flutter/material.dart';
-import 'package:job_card/route_generator.dart';
+// import 'package:job_card/route_generator.dart';
+import './screens_new/login.dart';
+import 'screens_new/main_screen.dart';
 
 void main() => runApp(MyApp());
 
@@ -7,21 +11,22 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      theme: ThemeData(
-        inputDecorationTheme: InputDecorationTheme(),
-        floatingActionButtonTheme:
-            FloatingActionButtonThemeData(backgroundColor: Color(0xFFF4442E)),
-        appBarTheme: AppBarTheme(
-          color: Color(0xFFF4442E),
-          iconTheme: IconThemeData(color: Colors.black),
-          textTheme: TextTheme(
-            title: TextStyle(color: Colors.black),
+        debugShowCheckedModeBanner: false,
+        theme: ThemeData(
+          inputDecorationTheme: InputDecorationTheme(),
+          floatingActionButtonTheme:
+              FloatingActionButtonThemeData(backgroundColor: Color(0xFFF4442E)),
+          appBarTheme: AppBarTheme(
+            color: Color(0xFFF4442E),
+            iconTheme: IconThemeData(color: Colors.black),
+            textTheme: TextTheme(
+              headline1: TextStyle(color: Colors.black),
+            ),
           ),
         ),
-      ),
-      onGenerateRoute: RouterGenerator.generateRoute,
-      initialRoute: '/resource_page',
-    );
+        home: JobPage()
+        // onGenerateRoute: RouterGenerator.generateRoute,
+        // initialRoute: '/resource_page',
+        );
   }
 }
