@@ -1,72 +1,72 @@
-import 'package:flutter/material.dart';
-import 'package:job_card/global_keys.dart';
-import 'package:job_card/screens/job_card_tab/approved_page.dart';
-import 'package:job_card/screens/job_card_tab/detailed_job_card_page.dart';
-import 'package:job_card/screens/job_card_tab/in_progress_page.dart';
-import 'package:job_card/screens/job_card_tab/my_job_card_home_page.dart';
-import 'package:job_card/screens/job_card_tab/not_started_page.dart';
-import 'package:job_card/screens/job_card_tab/waiting_for_approval_page.dart';
+// import 'package:flutter/material.dart';
+// import 'package:job_card/global_keys.dart';
+// import 'package:job_card/screens/job_card_tab/approved_page.dart';
+// import 'package:job_card/screens/job_card_tab/detailed_job_card_page.dart';
+// import 'package:job_card/screens/job_card_tab/in_progress_page.dart';
+// import 'package:job_card/screens/job_card_tab/my_job_card_home_page.dart';
+// import 'package:job_card/screens/job_card_tab/not_started_page.dart';
+// import 'package:job_card/screens/job_card_tab/waiting_for_approval_page.dart';
 
-class MyJobCardRoot extends StatefulWidget {
-  // MyJobCardRootPage({Key myJobCardRootPageKey}) : super(key: myJobCardRootPageKey);
-  @override
-  _MyJobCardRootState createState() => _MyJobCardRootState();
-}
+// class MyJobCardRoot extends StatefulWidget {
+//   // MyJobCardRootPage({Key myJobCardRootPageKey}) : super(key: myJobCardRootPageKey);
+//   @override
+//   _MyJobCardRootState createState() => _MyJobCardRootState();
+// }
 
-class _MyJobCardRootState extends State<MyJobCardRoot> {
-  @override
-  Widget build(BuildContext context) {
-    return Navigator(
-      onGenerateRoute: RouterGeneratorForJobCard.generateRoute,
-      key: myJobCardRootPageKey,
-      initialRoute: '/',
-    );
-  }
-}
+// class _MyJobCardRootState extends State<MyJobCardRoot> {
+//   @override
+//   Widget build(BuildContext context) {
+//     return Navigator(
+//       onGenerateRoute: RouterGeneratorForJobCard.generateRoute,
+//       key: myJobCardRootPageKey,
+//       initialRoute: '/',
+//     );
+//   }
+// }
 
-class RouterGeneratorForJobCard {
-  static Route<dynamic> generateRoute(RouteSettings settings) {
-    final args = settings.arguments;
-    print(settings.name);
-    switch (settings.name) {
-      case '/':
-        return MaterialPageRoute(
-            builder: (_) => MyJobCardsHomePage(), settings: settings);
+// class RouterGeneratorForJobCard {
+//   static Route<dynamic> generateRoute(RouteSettings settings) {
+//     final args = settings.arguments;
+//     print(settings.name);
+//     switch (settings.name) {
+//       case '/':
+//         return MaterialPageRoute(
+//             builder: (_) => MyJobCardsHomePage(), settings: settings);
 
-      case '/not_started_page':
-        return MaterialPageRoute(
-            builder: (_) => NotStartedPage(), settings: settings);
+//       case '/not_started_page':
+//         return MaterialPageRoute(
+//             builder: (_) => NotStartedPage(), settings: settings);
 
-      case '/in_progress_page':
-        return MaterialPageRoute(
-            builder: (_) => InProgressPage(), settings: settings);
+//       case '/in_progress_page':
+//         return MaterialPageRoute(
+//             builder: (_) => InProgressPage(), settings: settings);
 
-      case '/waiting_for_approval_page':
-        return MaterialPageRoute(
-            builder: (_) => WaitingForApprovalPage(), settings: settings);
+//       case '/waiting_for_approval_page':
+//         return MaterialPageRoute(
+//             builder: (_) => WaitingForApprovalPage(), settings: settings);
 
-      case '/approved_page':
-        return MaterialPageRoute(
-            builder: (_) => ApprovedPage(), settings: settings);
+//       case '/approved_page':
+//         return MaterialPageRoute(
+//             builder: (_) => ApprovedPage(), settings: settings);
 
-      case '/detailed_job_card':
-        return MaterialPageRoute(
-          builder: (_) => DetailedJobCardPage(
-            myJobCard: args,
-          ),
-        );
+//       case '/detailed_job_card':
+//         return MaterialPageRoute(
+//           builder: (_) => DetailedJobCardPage(
+//             myJobCard: args,
+//           ),
+//         );
 
-      default:
-        return errorRoute();
-    }
-  }
+//       default:
+//         return errorRoute();
+//     }
+//   }
 
-  static Route<dynamic> errorRoute() {
-    return MaterialPageRoute(builder: (_) {
-      return Scaffold(
-          body: Center(
-        child: Text('Error from jc root page'),
-      ));
-    });
-  }
-}
+//   static Route<dynamic> errorRoute() {
+//     return MaterialPageRoute(builder: (_) {
+//       return Scaffold(
+//           body: Center(
+//         child: Text('Error from jc root page'),
+//       ));
+//     });
+//   }
+// }
